requirements/development.txt
-r base.txt

pre-commit==3.7.1
ruff==0.5.4
black==24.4.2
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
django-cors-headers==4.6.0
